Isolate toolbox install and fix phar.xml path
Slightly refactors the env vars for clarity

Sets WORKDIR earlier - phive was creating `/.phive/phar.xml` outside toolbox dir

Isolates the toolbox-install stage, this means the last stage can do `COPY --from`. The impact there is that if the toolbox install created the same file structure, the existing layers can be reused. On local this isn't too significant but in a build-and-push the registry can use the layer it already has